+++ src/main/java/org/iesalixar/daw2/joseortega/dwese_ticket_logger_webapp/controllers/ProvinceController.java
@@ -8,6 +8,10 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.MessageSource;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
@@ -155,11 +159,28 @@ public class ProvinceController {
      * @param redirectAttributes Atributos para mensajes flash de redirección.
      * @return Redirección a la lista de provincias.
      */
+    @PreAuthorize("hasRole('ADMIN')")
     @PostMapping("/delete")
     public String deleteProvince(@RequestParam("id") Long id, RedirectAttributes redirectAttributes) {
         logger.info("Eliminando provincia con ID {}", id);
-        provinceRepository.deleteById(id);
-        logger.info("Provincia con ID {} eliminada con éxito.", id);
+        //Obtener el objeto de autenticación
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        // Verificar si el usuario tiene el rol ADMIN
+        if (auth == null || !auth.getAuthorities().contains(new SimpleGrantedAuthority("ROLE_ADMIN"))) {
+            String username = (auth != null) ? auth.getName() : "Usuario desconocido";
+            String errorMessage = "El usuario " + username + " no tiene permisos para borrar la provincia.";
+            logger.warn(errorMessage);
+            redirectAttributes.addFlashAttribute("errorMessage", errorMessage);
+            return "redirect:/403"; // Redirige a la página de error 403 o a otra página de tu elección
+        }
+        try {
+            provinceRepository.deleteById(id);
+            logger.info("Provincia con ID {} eliminada con éxito.", id);
+        } catch (Exception e) {
+            logger.error("Error al eliminar la provincia con ID {}: {}", id,
+                    e.getMessage());
+            redirectAttributes.addFlashAttribute("errorMessage", "Error al eliminar la provincia.");
+        }
         return "redirect:/provinces"; // Redirigir a la lista de provincias
     }
 }
