+++ src/main/java/org/iesalixar/daw2/joseortega/dwese_ticket_logger_webapp/controllers/SupermarketController.java
@@ -7,6 +7,10 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.MessageSource;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
@@ -142,11 +146,28 @@ public class SupermarketController {
      * @param redirectAttributes Atributos para mensajes flash de redirección.
      * @return Redirección a la lista de supermercados.
      */
+    @PreAuthorize("hasRole('ADMIN')")
     @PostMapping("/delete")
     public String deleteSupermarket(@RequestParam("id") Long id, RedirectAttributes redirectAttributes) {
         logger.info("Eliminando supermercado con ID {}", id);
-        supermarketRepository.deleteById(id);
-        logger.info("Supermercado con ID {} eliminada con éxito.", id);
-        return "redirect:/supermarkets"; // Redirigir a la lista de supermercados
+        //Obtener el objeto de autenticación
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        // Verificar si el usuario tiene el rol ADMIN
+        if (auth == null || !auth.getAuthorities().contains(new SimpleGrantedAuthority("ROLE_ADMIN"))) {
+            String username = (auth != null) ? auth.getName() : "Usuario desconocido";
+            String errorMessage = "El usuario " + username + " no tiene permisos para borrar el supermercado.";
+            logger.warn(errorMessage);
+            redirectAttributes.addFlashAttribute("errorMessage", errorMessage);
+            return "redirect:/403"; // Redirige a la página de error 403 o a otra página de tu elección
+        }
+        try {
+            supermarketRepository.deleteById(id);
+            logger.info("Supermercado con ID {} eliminada con éxito.", id);
+        }catch(Exception e) {
+            logger.error("Error al eliminar el supermercado con ID {}: {}", id,
+                    e.getMessage());
+            redirectAttributes.addFlashAttribute("errorMessage", "Error al eliminar supermercado.");
+        }
+        return "redirect:/supermarkets"; // Redirigir a la lista de categorias
     }
 }
